backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest-django==4.7.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
coreapi==2.3.3 